Разделен view controller сканера и capture session
